semantics: Update generator update interface
Currently generators do not get updated with the non-deterministic input chosen
by the system from the next macrostep. This limits the modeling capabilities of
our generators. To support this feature, the update interface of the generator
must be extended and it has to be used after the non-deterministic input choice
is made.

- Add extra input choice argument to update the update interface of the
  generator.

- Change invocation of the update interface to right after the non-determisitic
  input choice.