semantics: Extend sequence generator
Currently inputs are generated according to the specifictaion in the script.

The official executive appends empty inputs when the environment does not provide
inputs but there are nodes that still can transition in order to generate empty
macrosteps, which allows the nodes that can transition to do so.

To support this feature, the input generator must be updated so that it considers
the state of every node in the system and appends empty inputs when any node has
not reached the state 'finished'.

- Add extra case in input generator to check for the state of all nodes and
generate an empty input.